# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "D&D Backend API",
        "description": "API para gerenciamento de campanhas de Dungeons & Dragons",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server v1"
        }
    ],
    "paths": {
...
        "/game-sessions/{id}/logs": {
            "get": {
                "tags": [
                    "Game Sessions"
                ],
                "summary": "Obter histórico de combate",
                "description": "Get combat logs",
                "operationId": "b4fc2e7e7b7a958a3fd7096f231bf86f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da sessão de jogo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Histórico de combate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "logs": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Jogador atacou causando 15 de dano!"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "request_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não autenticado."
                                        },
                                        "meta": {
                                            "properties": {
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "request_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Esta ação não é autorizada."
                                        },
                                        "meta": {
                                            "properties": {
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "request_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sessão não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recurso não encontrado."
                                        },
                                        "meta": {
                                            "properties": {
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "request_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}